Refactor load_raw_set
Make it more general. The subjects to be loaded are now
specified with the --subj-ids argument. The file name pattern
changes to subj-<id>.mat, and that file is supposed to have some
variables that were previously computed in Python. Although this is
a regression (we want to move everything to Python), this simplifies
the code, for now. We want later to generate the noisy ICLabel labels
with MNE_ICALabel instead of the EEGLAB Matlab plugin.