Calibrate idf for cue dataset
We found that the distribution of bowav rates in cue is very different than in
emotion_study. This calibration aims to align the mean tf-idf values in the
source (emotion_study) and the target (cue), allowing up to a 10-fold deviation
of the target bowav rates from the source bowav rates.